# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minutor API",
        "description": "Sistema de apontamento de horas e despesas",
        "contact": {
            "email": "admin@minutor.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Servidor de Desenvolvimento"
        }
    ],
    "paths": {
...
        "/api/v1/users/{user}/roles": {
            "get": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Listar roles do usuário",
                "description": "Lista todos os roles atribuídos a um usuário",
                "operationId": "6b410cce0002404172d0d96de07fa053",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Roles do usuário",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "user": {
                                                    "type": "object"
                                                },
                                                "roles": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Sincronizar roles do usuário",
                "description": "Substitui todos os roles do usuário pelos especificados",
                "operationId": "c3202824b96e9150caf0117e3af7d346",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roles"
                                ],
                                "properties": {
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Project Manager"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Roles sincronizados com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Atribuir roles ao usuário",
                "description": "Adiciona um ou mais roles a um usuário",
                "operationId": "aa483090a1a0be2bc6a052a25178c86c",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roles"
                                ],
                                "properties": {
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Administrator",
                                            "Project Manager"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Roles atribuídos com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Remover roles do usuário",
                "description": "Remove um ou mais roles de um usuário",
                "operationId": "acbb76b43d770f21719dd6ed3b7d3a62",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "roles"
                                ],
                                "properties": {
                                    "roles": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Consultant"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Roles removidos com sucesso"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/permissions": {
            "get": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Listar permissões do usuário",
                "description": "Lista todas as permissões do usuário (via roles e diretas)",
                "operationId": "e8456c4b5bae1a2e16635abde775080d",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Permissões do usuário"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/users/{user}/permissions/check": {
            "post": {
                "tags": [
                    "User Roles"
                ],
                "summary": "Verificar permissão do usuário",
                "description": "Verifica se o usuário tem uma permissão específica",
                "operationId": "284c22948ee16ff215bb58d981f8e883",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "ID do usuário",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "permission"
                                ],
                                "properties": {
                                    "permission": {
                                        "type": "string",
                                        "example": "projects.create"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Resultado da verificação de permissão"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Token de autenticação Bearer obtido através do endpoint de login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}